# api/openapi.yml
openapi: 3.0.3
info:
  title: Change Quality Agent API
  version: 0.1.0
  description: >
    API contract for SOP quality runs, run history, and generic run observation.
    SOP fetching is mocked in v1.
servers:
  - url: /
tags:
  - name: sop
    description: SOP entry points and SOP-scoped run history.
  - name: runs
    description: Generic run observation APIs.
  - name: mcp
    description: MCP server configuration and runtime management APIs.
paths:
  /api/sop/environments:
    get:
      tags:
        - sop
      summary: List configured SOP environments
      operationId: listSopEnvironments
      responses:
        "200":
          description: Configured public environments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnvironmentPublic"
  /api/sop/{sop_id}:
    get:
      tags:
        - sop
      summary: Preview a mocked SOP snapshot
      operationId: getSopPreview
      parameters:
        - $ref: "#/components/parameters/SopId"
        - $ref: "#/components/parameters/EnvironmentKey"
      responses:
        "200":
          description: Mocked SOP snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SopSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/sop/{sop_id}/runs:
    post:
      tags:
        - sop
      summary: Start a quality run for a mocked SOP snapshot
      operationId: startSopRun
      parameters:
        - $ref: "#/components/parameters/SopId"
        - $ref: "#/components/parameters/EnvironmentKey"
      responses:
        "202":
          description: Run accepted and scheduled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunStartResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: An active run already exists for this SOP and environment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveRunConflict"
        "502":
          $ref: "#/components/responses/BadGateway"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      tags:
        - sop
      summary: List SOP run history
      operationId: listSopRuns
      parameters:
        - $ref: "#/components/parameters/SopId"
        - $ref: "#/components/parameters/EnvironmentKey"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: SOP run history, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RunSummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/sop/recent/runs:
    get:
      tags:
        - sop
      summary: List recent SOP runs for one environment
      operationId: listRecentSopRuns
      parameters:
        - $ref: "#/components/parameters/EnvironmentKey"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Recent SOP runs for the environment, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RunSummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/runs/{run_id}:
    get:
      tags:
        - runs
      summary: Get generic run details
      description: >
        Generic run details intentionally do not expose SOP-specific fields such
        as env_key at the top level.
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: debug
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Generic run details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/runs/{run_id}/events:
    get:
      tags:
        - runs
      summary: Stream persisted run events
      description: >
        Server-Sent Events stream. The server first replays persisted events
        with sequence greater than after, then polls persisted events until it
        sends terminal done or error.
      operationId: streamRunEvents
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: after
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: >
            SSE frames using id as sequence, event as event type, and data as a
            JSON RunEvent envelope.
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  id: 3
                  event: done
                  data: {"run_id":"2f4f9c1c-5f97-4c9c-89d2-9e7e3b6a1a11","sequence":3,"type":"done","node":null,"thread_id":"thread-id","checkpoint_id":null,"task_id":null,"payload":{"status":"done"},"created_at":"2026-05-25T00:00:00Z"}
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/mcp/servers:
    get:
      tags:
        - mcp
      summary: List MCP servers
      operationId: listMcpServers
      security:
        - McpAdminToken: []
      responses:
        "200":
          description: MCP servers with redacted config values.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McpServerSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      tags:
        - mcp
      summary: Create an MCP server configuration
      operationId: createMcpServer
      security:
        - McpAdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpServerCreate"
      responses:
        "201":
          description: Created MCP server with redacted config values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServerDetail"
        "409":
          description: An MCP server with this name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/mcp/servers/{server_id}:
    get:
      tags:
        - mcp
      summary: Get MCP server details
      operationId: getMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      responses:
        "200":
          description: MCP server details and latest tool snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServerDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
    patch:
      tags:
        - mcp
      summary: Update an MCP server configuration
      operationId: updateMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McpServerUpdate"
      responses:
        "200":
          description: Updated MCP server with redacted config values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServerDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Running MCP servers must be stopped before config updates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags:
        - mcp
      summary: Delete an MCP server configuration
      operationId: deleteMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      responses:
        "204":
          description: MCP server deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/mcp/servers/{server_id}/start:
    post:
      tags:
        - mcp
      summary: Start an MCP server
      operationId: startMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      responses:
        "200":
          description: Runtime status after start attempt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpLifecycleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/mcp/servers/{server_id}/stop:
    post:
      tags:
        - mcp
      summary: Stop an MCP server
      operationId: stopMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      responses:
        "200":
          description: Runtime status after stop attempt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpLifecycleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/mcp/servers/{server_id}/restart:
    post:
      tags:
        - mcp
      summary: Restart an MCP server
      operationId: restartMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      responses:
        "200":
          description: Runtime status after restart attempt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpLifecycleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/mcp/servers/{server_id}/check:
    post:
      tags:
        - mcp
      summary: Check an MCP server and refresh tools
      operationId: checkMcpServer
      security:
        - McpAdminToken: []
      parameters:
        - $ref: "#/components/parameters/McpServerId"
      responses:
        "200":
          description: Runtime status after check attempt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpLifecycleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "422":
          $ref: "#/components/responses/ValidationError"
components:
  securitySchemes:
    McpAdminToken:
      type: apiKey
      in: header
      name: X-MCP-Admin-Token
  parameters:
    SopId:
      name: sop_id
      in: path
      required: true
      schema:
        type: string
      example: release-checklist
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    McpServerId:
      name: server_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EnvironmentKey:
      name: env
      in: query
      required: true
      schema:
        type: string
      example: dev
  responses:
    NotFound:
      description: Unknown environment, SOP, or run.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream dependency or lifecycle failure before completion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Missing or invalid MCP admin token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: MCP runtime is not configured or not enabled.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
    EnvironmentPublic:
      type: object
      required:
        - key
        - name_zh
        - name_en
      properties:
        key:
          type: string
          example: dev
        name_zh:
          type: string
          example: 开发
        name_en:
          type: string
          example: Development
    SopSnapshot:
      type: object
      required:
        - sop_id
        - env_key
        - payload
      properties:
        sop_id:
          type: string
        env_key:
          type: string
        source_version:
          type: string
          nullable: true
          example: mock-v1
        updated_at:
          type: string
          format: date-time
          nullable: true
        payload:
          type: object
          additionalProperties: true
    RunStatus:
      type: string
      enum:
        - pending
        - running
        - success
        - error
        - timeout
        - interrupted
    RunStartResponse:
      type: object
      required:
        - run_id
        - status
        - status_url
        - events_url
      properties:
        run_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        status_url:
          type: string
        events_url:
          type: string
    ActiveRunConflict:
      type: object
      required:
        - message
        - active_run_id
        - status_url
        - events_url
      properties:
        message:
          type: string
        active_run_id:
          type: string
          format: uuid
        status_url:
          type: string
        events_url:
          type: string
    RunSummary:
      type: object
      required:
        - run_id
        - subject_type
        - subject_id
        - status
        - completed_nodes
        - latest_sequence
      properties:
        run_id:
          type: string
          format: uuid
        subject_type:
          type: string
          example: sop
        subject_id:
          type: string
          example: release-checklist
        status:
          $ref: "#/components/schemas/RunStatus"
        current_node:
          type: string
          nullable: true
        completed_nodes:
          type: array
          items:
            type: string
        latest_sequence:
          type: integer
          minimum: 0
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        result_status:
          type: string
          nullable: true
        error_summary:
          type: string
          nullable: true
    RunDebug:
      type: object
      required:
        - thread_id
      properties:
        thread_id:
          type: string
        current_checkpoint_id:
          type: string
          nullable: true
        langgraph_state_snapshot:
          type: object
          nullable: true
          additionalProperties: true
        raw_graph_output:
          type: object
          nullable: true
          additionalProperties: true
        raw_last_event:
          allOf:
            - $ref: "#/components/schemas/RunEvent"
          nullable: true
    RunDetail:
      allOf:
        - $ref: "#/components/schemas/RunSummary"
        - type: object
          required:
            - debug
          properties:
            debug:
              allOf:
                - $ref: "#/components/schemas/RunDebug"
              nullable: true
    RunEventType:
      type: string
      enum:
        - tasks
        - messages
        - updates
        - custom
        - checkpoints
        - error
        - done
    RunEvent:
      type: object
      required:
        - run_id
        - sequence
        - type
        - thread_id
        - payload
        - created_at
      properties:
        run_id:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 1
        type:
          $ref: "#/components/schemas/RunEventType"
        node:
          type: string
          nullable: true
        thread_id:
          type: string
        checkpoint_id:
          type: string
          nullable: true
        task_id:
          type: string
          nullable: true
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    McpTransport:
      type: string
      enum:
        - stdio
        - http
    McpDesiredState:
      type: string
      enum:
        - running
        - stopped
    McpRuntimeStatus:
      type: string
      enum:
        - unknown
        - starting
        - running
        - stopping
        - stopped
        - error
    McpServerCreate:
      type: object
      required:
        - name
        - transport
      properties:
        name:
          type: string
          example: filesystem
        transport:
          $ref: "#/components/schemas/McpTransport"
        command:
          type: string
          nullable: true
          example: uvx
        args:
          type: array
          items:
            type: string
          default: []
          example:
            - mcp-server-filesystem
        env:
          type: object
          additionalProperties:
            type: string
          default: {}
          example:
            TOKEN: secret
        url:
          type: string
          nullable: true
          example: https://example.com/mcp
        headers:
          type: object
          additionalProperties:
            type: string
          default: {}
        enabled:
          type: boolean
          default: false
        desired_state:
          allOf:
            - $ref: "#/components/schemas/McpDesiredState"
          default: stopped
    McpServerUpdate:
      type: object
      properties:
        name:
          type: string
        command:
          type: string
          nullable: true
        args:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: string
        url:
          type: string
          nullable: true
        headers:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
        desired_state:
          $ref: "#/components/schemas/McpDesiredState"
    McpServerTool:
      type: object
      required:
        - name
        - input_schema
      properties:
        name:
          type: string
          example: search
        description:
          type: string
          nullable: true
        input_schema:
          type: object
          additionalProperties: true
          default: {}
        discovered_at:
          type: string
          format: date-time
          nullable: true
    McpServerSummary:
      type: object
      required:
        - id
        - name
        - transport
        - args
        - env
        - headers
        - enabled
        - desired_state
        - runtime_status
        - tool_count
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        transport:
          $ref: "#/components/schemas/McpTransport"
        command:
          type: string
          nullable: true
        args:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: string
          example:
            TOKEN: "********"
        url:
          type: string
          nullable: true
        headers:
          type: object
          additionalProperties:
            type: string
          example:
            Authorization: "********"
        enabled:
          type: boolean
        desired_state:
          $ref: "#/components/schemas/McpDesiredState"
        runtime_status:
          $ref: "#/components/schemas/McpRuntimeStatus"
        last_checked_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
        tool_count:
          type: integer
          minimum: 0
    McpServerDetail:
      allOf:
        - $ref: "#/components/schemas/McpServerSummary"
        - type: object
          required:
            - tools
          properties:
            tools:
              type: array
              items:
                $ref: "#/components/schemas/McpServerTool"
    McpLifecycleResponse:
      type: object
      required:
        - server_id
        - desired_state
        - runtime_status
        - tool_count
      properties:
        server_id:
          type: string
          format: uuid
        desired_state:
          $ref: "#/components/schemas/McpDesiredState"
        runtime_status:
          $ref: "#/components/schemas/McpRuntimeStatus"
        last_checked_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
        tool_count:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            additionalProperties: true